refactor: 테스트 코드 중복 제거
