+++ src/test/java/subway/application/LineServiceTest.java
@@ -3,9 +3,6 @@ package subway.application;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
 import subway.application.dto.LineDto;
 import subway.domain.*;
 
@@ -13,10 +10,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ActiveProfiles("test")
-@Sql("/initialization.sql")
-@SpringBootTest
-class LineServiceTest {
+class LineServiceTest extends ServiceTest {
     @Autowired
     private LineService lineService;
