+++ src/test/java/subway/application/PathServiceTest.java
@@ -3,9 +3,6 @@ package subway.application;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
 import subway.application.dto.PathDto;
 import subway.domain.*;
 
@@ -14,10 +11,7 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static subway.application.PathServiceTest.Data.*;
 
-@ActiveProfiles("test")
-@Sql("/initialization.sql")
-@SpringBootTest
-class PathServiceTest {
+class PathServiceTest extends ServiceTest{
     @Autowired
     private LineRepository lineRepository;
