+++ src/test/java/subway/application/SectionServiceTest.java
@@ -3,18 +3,12 @@ package subway.application;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
 import subway.application.dto.SectionDto;
 import subway.domain.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ActiveProfiles("test")
-@Sql("/initialization.sql")
-@SpringBootTest
-class SectionServiceTest {
+class SectionServiceTest extends ServiceTest{
     @Autowired
     private SectionService sectionService;
