+++ src/test/java/subway/application/ServiceTest.java
@@ -1,6 +1,5 @@
-package subway;
+package subway.application;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
@@ -8,10 +7,5 @@ import org.springframework.test.context.jdbc.Sql;
 @ActiveProfiles("test")
 @Sql("/initialization.sql")
 @SpringBootTest
-class SubwayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
+public class ServiceTest {
 }
