+++ src/test/java/subway/integration/IntegrationTest.java
@@ -1,0 +1,15 @@
+package subway.integration;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.jdbc.Sql;
+
+@ActiveProfiles("test")
+@Sql("/initialization.sql")
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class IntegrationTest {
+    @LocalServerPort
+    int port;
+
+}
