+++ src/test/java/subway/integration/LineIntegrationTest.java
@@ -6,12 +6,8 @@ import io.restassured.response.Response;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
 import subway.presentation.dto.request.LineRequest;
 import subway.presentation.dto.response.LineResponse;
 
@@ -21,20 +17,14 @@ import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ActiveProfiles("test")
-@Sql("/initialization.sql")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DisplayName("지하철 노선 관련 기능")
-public class LineIntegrationTest {
+public class LineIntegrationTest extends IntegrationTest {
     private static final LineRequest lineRequest1 = new LineRequest("신분당선");
     private static final LineRequest lineRequest2 = new LineRequest("구신분당선");
 
-    @LocalServerPort
-    private int port;
-
     @BeforeEach
     public void setUp() {
-        RestAssured.port = port;
+        RestAssured.port = super.port;
     }
 
     @DisplayName("지하철 노선을 생성한다.")
